temp

+++ network/cvae_graph/dataloader.py
@@ -112,7 +112,7 @@ class GraphDataset(Dataset):
             building_masks = np.array([graph.nodes[node]['building_masks'] for node in graph.nodes()], dtype=object)
 
             # PyTorch 텐서로 변환 (필요한 경우)
-            building_masks = torch.tensor(list(building_masks), dtype=torch.float32)
+            building_masks = torch.tensor(list(building_masks), dtype=torch.long)
 
             if self.condition_type == 'image' or self.condition_type == 'image_resnet34':
                 condition = torch.tensor(np.array(graph.graph['condition']), dtype=torch.float32)
